More V&V, unit test, reconcile peer review
I reqquest a technical peer review from Abram Claycomb, now at the National
Oceanic and Atmospheric Administration (NOAA). He looked through the code,
"kicked the tires", and could find no obvious technical problems with the
software. His emailed response was in fact:

>   Josh,
>
>   Actually, I was trying to come up with some constructive criticisms or
>   stuff I didn't understand.  I thought I would break it or I'd have
>   difficulty on the noaa network, but no problems.  So, I want to say
>   well done, it installs, connects, downloads, works and outputs seemingly
>   flawlessly.  If there are specific questions you had answers to that you
>   want asked, I can do that, but otherwise I'll open an 'issue' and say
>   well done, no technical questions from Abe.
>
>   Cheers,
>
>   Abe

Since Abe asked for so much, I went ahead and expanded the V&V results
presented in the SECS_Verify_Demo.ipynb notebook, tweaked the README,
regenerated output, rearranged the package directory hiearchy a bit,
and created some actual Nose-friendly unit tests.

The package is now ready for a formal administrative code review. If
you are so fortunate as to be the reviewer, please add comments to the
pull request that will get submited with this commit. -EJR